requirements.txt
altgraph==0.17
astroid==2.4.1
autopep8==1.5.2
colorama==0.4.3
comtypes==1.1.7
future==0.18.2
isort==4.3.21
keyboard==0.13.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
MouseInfo==0.1.3
numpy==1.18.4
pefile==2019.4.18
Pillow==7.1.2
PyAutoGUI==0.9.50
pycodestyle==2.6.0
PyGetWindow==0.0.8
PyInstaller==3.6
pylint==2.5.2
PyMsgBox==1.0.8
pynput==1.6.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
pystray==0.15.0
python-dotenv==0.13.0
PyTweening==1.0.3
pywin32==227
pywin32-ctypes==0.2.0
pywinauto==0.6.8
six==1.14.0
toml==0.10.1
wrapt==1.12.1
wxPython==4.1.0
